+++ imgaug/imgaug.py
@@ -812,8 +812,10 @@ def angle_between_vectors(v1, v2):
     3.141592...
 
     """
-    v1_u = v1 / np.linalg.norm(v1)
-    v2_u = v2 / np.linalg.norm(v2)
+    l1 = np.linalg.norm(v1)
+    l2 = np.linalg.norm(v2)
+    v1_u = (v1 / l1) if l1 > 0 else np.float32(v1) * 0
+    v2_u = (v2 / l2) if l2 > 0 else np.float32(v2) * 0
     return np.arccos(np.clip(np.dot(v1_u, v2_u), -1.0, 1.0))
 
 
@@ -4298,7 +4300,7 @@ class Polygon(object):
         Returns
         -------
         imgaug.BoundingBox
-            The bounding box tightly containing the polygon.
+            Tight bounding box around the polygon.
 
         """
         xx = self.xx
@@ -4577,6 +4579,461 @@ def _interpolate_points_by_max_distance(points, max_distance, closed=True):
     if not closed:
         points_interp.append(points[-1])
     return points_interp
+
+
+class _ConcavePolygonRecoverer(object):
+    def __init__(self, threshold_duplicate_points=1e-4, noise_strength=1e-4,
+                 oversampling=0.01, max_segment_difference=1e-4):
+        self.threshold_duplicate_points = threshold_duplicate_points
+        self.noise_strength = noise_strength
+        self.oversampling = oversampling
+        self.max_segment_difference = max_segment_difference
+
+        # this limits the maximum amount of points after oversampling, i.e.
+        # if N points are input into oversampling, then M oversampled points are
+        # generated such that N+M <= this value
+        self.oversample_up_to_n_points_max = 75
+
+        # ----
+        # parameters for _fit_best_valid_polygon()
+        # ----
+        # how many changes may be done max to the initial (convex hull) polygon
+        # before simply returning the result
+        self.fit_n_changes_max = 100
+        # for how many iterations the optimization loop may run max
+        # before simply returning the result
+        self.fit_n_iters_max = 3
+        # how far (wrt. to their position in the input list) two points may be
+        # apart max to consider adding an edge between them (in the first loop
+        # iteration and the ones after that)
+        self.fit_max_dist_first_iter = 1
+        self.fit_max_dist_other_iters = 2
+        # The fit loop first generates candidate edges and then modifies the
+        # polygon based on these candidates. This limits the maximum amount
+        # of considered candidates. If the number is less than the possible
+        # number of candidates, they are randomly subsampled. Values beyond
+        # 100 significantly increase runtime (for polygons that reach that
+        # number).
+        self.fit_n_candidates_before_sort_max = 100
+
+    def recover_from(self, new_exterior, old_polygon, random_state=0):
+        assert isinstance(new_exterior, list) or (
+                is_np_array(new_exterior)
+                and new_exterior.ndim == 2
+                and new_exterior.shape[1] == 2)
+        assert len(new_exterior) >= 3, \
+            "Cannot recover a concave polygon from less than three points."
+
+        # create Polygon instance, if it is already valid then just return
+        # immediately
+        polygon = old_polygon.deepcopy(exterior=new_exterior)
+        if polygon.is_valid:
+            return polygon
+
+        if not isinstance(random_state, np.random.RandomState):
+            random_state = np.random.RandomState(random_state)
+        rss = derive_random_states(random_state, 3)
+
+        # remove consecutive duplicate points
+        new_exterior = self._remove_consecutive_duplicate_points(new_exterior)
+
+        # check that points are not all identical or on a line
+        new_exterior = self._fix_polygon_is_line(new_exterior, rss[0])
+
+        # jitter duplicate points
+        new_exterior = self._jitter_duplicate_points(new_exterior, rss[1])
+
+        # generate intersection points
+        segment_add_points = self._generate_intersection_points(new_exterior)
+
+        # oversample points around intersections
+        if self.oversampling is not None and self.oversampling > 0:
+            segment_add_points = self._oversample_intersection_points(
+                new_exterior, segment_add_points)
+
+        # integrate new points into exterior
+        new_exterior_inter = self._insert_intersection_points(
+            new_exterior, segment_add_points)
+
+        # find best fit polygon, starting from convext polygon
+        new_exterior_concave_ids = self._fit_best_valid_polygon(new_exterior_inter, rss[2])
+        new_exterior_concave = [new_exterior_inter[idx] for idx in new_exterior_concave_ids]
+
+        return old_polygon.deepcopy(exterior=new_exterior_concave)
+
+    def _remove_consecutive_duplicate_points(self, points):
+        result = []
+        for point in points:
+            if result:
+                dist = np.linalg.norm(np.float32(point) - np.float32(result[-1]))
+                is_same = (dist < self.threshold_duplicate_points)
+                if not is_same:
+                    result.append(point)
+            else:
+                result.append(point)
+        if len(result) >= 2:
+            dist = np.linalg.norm(np.float32(result[0]) - np.float32(result[-1]))
+            is_same = (dist < self.threshold_duplicate_points)
+            result = result[0:-1] if is_same else result
+        return result
+
+    # fix polygons for which all points are on a line
+    def _fix_polygon_is_line(self, exterior, random_state):
+        assert len(exterior) >= 3
+        noise_strength = self.noise_strength
+        while self._is_polygon_line(exterior):
+            noise = random_state.uniform(
+                -noise_strength, noise_strength, size=(len(exterior), 2)
+            ).astype(np.float32)
+            exterior = [(point[0] + noise_i[0], point[1] + noise_i[1])
+                        for point, noise_i in zip(exterior, noise)]
+            noise_strength = noise_strength * 10
+            assert noise_strength > 0
+        return exterior
+
+    @classmethod
+    def _is_polygon_line(cls, exterior):
+        vec_down = np.float32([0, 1])
+        p1 = exterior[0]
+        angles = set()
+        for p2 in exterior[1:]:
+            vec = np.float32(p2) - np.float32(p1)
+            angle = angle_between_vectors(vec_down, vec)
+            angles.add(int(angle * 1000))
+        return len(angles) <= 1
+
+    def _jitter_duplicate_points(self, exterior, random_state):
+        def _find_duplicates(exterior_with_duplicates):
+            points_map = collections.defaultdict(list)
+
+            for i, point in enumerate(exterior_with_duplicates):
+                # we use 10/x here to be a bit more lenient, the precise
+                # distance test is further below
+                x = int(np.round(point[0] * ((1/10) / self.threshold_duplicate_points)))
+                y = int(np.round(point[1] * ((1/10) / self.threshold_duplicate_points)))
+                for d0 in [-1, 0, 1]:
+                    for d1 in [-1, 0, 1]:
+                        points_map[(x+d0, y+d1)].append(i)
+
+            duplicates = [False] * len(exterior_with_duplicates)
+            for key in points_map:
+                candidates = points_map[key]
+                for i in range(len(candidates)):
+                    p0_idx = candidates[i]
+                    p0 = exterior_with_duplicates[p0_idx]
+                    if duplicates[p0_idx]:
+                        continue
+
+                    for j in range(i+1, len(candidates)):
+                        p1_idx = candidates[j]
+                        p1 = exterior_with_duplicates[p1_idx]
+                        if duplicates[p1_idx]:
+                            continue
+
+                        dist = np.sqrt((p0[0] - p1[0])**2 + (p0[1] - p1[1])**2)
+                        if dist < self.threshold_duplicate_points:
+                            duplicates[p1_idx] = True
+
+            return duplicates
+
+        noise_strength = self.noise_strength
+        assert noise_strength > 0
+        exterior = exterior[:]
+        converged = False
+        while not converged:
+            duplicates = _find_duplicates(exterior)
+            if any(duplicates):
+                noise = random_state.uniform(
+                    -self.noise_strength, self.noise_strength, size=(len(exterior), 2)
+                ).astype(np.float32)
+                for i, is_duplicate in enumerate(duplicates):
+                    if is_duplicate:
+                        exterior[i] = (exterior[i][0] + noise[i][0], exterior[i][1] + noise[i][1])
+
+                noise_strength *= 10
+            else:
+                converged = True
+
+        return exterior
+
+    # TODO remove?
+    @classmethod
+    def _calculate_circumference(cls, points):
+        assert len(points) >= 3
+        points = np.array(points, dtype=np.float32)
+        points_matrix = np.zeros((len(points), 4), dtype=np.float32)
+        points_matrix[:, 0:2] = points
+        points_matrix[0:-1, 2:4] = points_matrix[1:, 0:2]
+        points_matrix[-1, 2:4] = points_matrix[0, 0:2]
+        distances = np.linalg.norm(
+            points_matrix[:, 0:2] - points_matrix[:, 2:4], axis=1)
+        return np.sum(distances)
+
+    def _generate_intersection_points(self, exterior, one_point_per_intersection=True):
+        assert isinstance(exterior, list)
+        assert all([len(point) == 2 for point in exterior])
+        if len(exterior) <= 0:
+            return []
+
+        # use (*[i][0], *[i][1]) formulation here imnstead of just *[i],
+        # because this way we convert numpy arrays to tuples of floats, which
+        # is required by isect_segments_include_segments
+        segments = [
+            (
+                (exterior[i][0], exterior[i][1]),
+                (exterior[(i + 1) % len(exterior)][0], exterior[(i + 1) % len(exterior)][1])
+            )
+            for i in range(len(exterior))
+        ]
+
+        # returns [(point, [(segment_p0, segment_p1), ..]), ...]
+        from imgaug.external.poly_point_isect_py2py3 import isect_segments_include_segments
+        intersections = isect_segments_include_segments(segments)
+
+        # estimate to which segment the found intersection points belong
+        segments_add_points = [[] for _ in range(len(segments))]
+        for point, associated_segments in intersections:
+            # the intersection point may be associated with multiple segments,
+            # but we only want to add it once, so pick the first segment
+            if one_point_per_intersection:
+                associated_segments = [associated_segments[0]]
+
+            for seg_inter_p0, seg_inter_p1 in associated_segments:
+                diffs = []
+                dists = []
+                for seg_p0, seg_p1 in segments:
+                    dist_p0p0 = np.linalg.norm(seg_p0 - np.array(seg_inter_p0))
+                    dist_p1p1 = np.linalg.norm(seg_p1 - np.array(seg_inter_p1))
+                    dist_p0p1 = np.linalg.norm(seg_p0 - np.array(seg_inter_p1))
+                    dist_p1p0 = np.linalg.norm(seg_p1 - np.array(seg_inter_p0))
+                    diff = min(dist_p0p0 + dist_p1p1, dist_p0p1 + dist_p1p0)
+                    diffs.append(diff)
+                    dists.append(np.linalg.norm(
+                        (seg_p0[0] - point[0], seg_p0[1] - point[1])
+                    ))
+
+                min_diff = np.min(diffs)
+                if min_diff < self.max_segment_difference:
+                    idx = int(np.argmin(diffs))
+                    segments_add_points[idx].append((point, dists[idx]))
+                else:
+                    warnings.warn(
+                        "Couldn't find fitting segment in "
+                        "_generate_intersection_points(). Ignoring intersection "
+                        "point.")
+
+        # sort intersection points by their distance to point 0 in each segment
+        # (clockwise ordering, this does something only for segments with
+        # >=2 intersection points)
+        segment_add_points_sorted = []
+        for idx in range(len(segments_add_points)):
+            points = [t[0] for t in segments_add_points[idx]]
+            dists = [t[1] for t in segments_add_points[idx]]
+            if len(points) < 2:
+                segment_add_points_sorted.append(points)
+            else:
+                both = sorted(zip(points, dists), key=lambda t: t[1])
+                # keep points, drop distances
+                segment_add_points_sorted.append([a for a, _b in both])
+        return segment_add_points_sorted
+
+    def _oversample_intersection_points(self, exterior, segment_add_points):
+        # segment_add_points must be sorted
+
+        if self.oversampling is None or self.oversampling <= 0:
+            return segment_add_points
+
+        segment_add_points_sorted_overs = [[] for _ in range(len(segment_add_points))]
+
+        n_points = len(exterior)
+        for i in range(len(exterior)):
+            last = exterior[i]
+            for j, p_inter in enumerate(segment_add_points[i]):
+                direction = (p_inter[0] - last[0], p_inter[1] - last[1])
+
+                if j == 0:
+                    # previous point was non-intersection, place 1 new point
+                    oversample = [1.0 - self.oversampling]
+                else:
+                    # previous point was intersection, place 2 new points
+                    oversample = [self.oversampling, 1.0 - self.oversampling]
+
+                for dist in oversample:
+                    point_over = (last[0] + dist * direction[0],
+                                  last[1] + dist * direction[1])
+                    segment_add_points_sorted_overs[i].append(point_over)
+                segment_add_points_sorted_overs[i].append(p_inter)
+                last = p_inter
+
+                is_last_in_group = (j == len(segment_add_points[i]) - 1)
+                if is_last_in_group:
+                    # previous point was oversampled, next point is
+                    # non-intersection, place 1 new point between the two
+                    exterior_point = exterior[(i + 1) % len(exterior)]
+                    direction = (exterior_point[0] - last[0],
+                                 exterior_point[1] - last[1])
+                    segment_add_points_sorted_overs[i].append(
+                        (last[0] + self.oversampling * direction[0],
+                         last[1] + self.oversampling * direction[1])
+                    )
+                    last = segment_add_points_sorted_overs[i][-1]
+
+                n_points += len(segment_add_points_sorted_overs[i])
+                if n_points > self.oversample_up_to_n_points_max:
+                    return segment_add_points_sorted_overs
+
+        return segment_add_points_sorted_overs
+
+    @classmethod
+    def _insert_intersection_points(cls, exterior, segment_add_points):
+        # segment_add_points must be sorted
+
+        assert len(exterior) == len(segment_add_points)
+        exterior_interp = []
+        for i in range(len(exterior)):
+            p0 = exterior[i]
+            exterior_interp.append(p0)
+            for j, p_inter in enumerate(segment_add_points[i]):
+                exterior_interp.append(p_inter)
+        return exterior_interp
+
+    def _fit_best_valid_polygon(self, points, random_state):
+        if len(points) < 2:
+            return None
+
+        def _compute_distance_point_to_line(point, line_start, line_end):
+            x_diff = line_end[0] - line_start[0]
+            y_diff = line_end[1] - line_start[1]
+            num = abs(
+                y_diff*point[0] - x_diff*point[1]
+                + line_end[0]*line_start[1] - line_end[1]*line_start[0]
+            )
+            den = np.sqrt(y_diff**2 + x_diff**2)
+            if den == 0:
+                return np.sqrt((point[0] - line_start[0])**2 + (point[1] - line_start[1])**2)
+            return num / den
+
+        poly = Polygon(points)
+        if poly.is_valid:
+            return sm.xrange(len(points))
+
+        hull = scipy.spatial.ConvexHull(points)
+        points_kept = list(hull.vertices)
+        points_left = [i for i in range(len(points)) if i not in hull.vertices]
+
+        iteration = 0
+        n_changes = 0
+        converged = False
+        while not converged:
+            candidates = []
+
+            # estimate distance metrics for points-segment pairs:
+            #  (1) distance (in vertices) between point and segment-start-point
+            #      in original input point chain
+            #  (2) euclidean distance between point and segment/line
+            # TODO this can be done more efficiently by caching the values and
+            #      only computing distances to segments that have changed in
+            #      the last iteration
+            # TODO these distances are not really the best metrics here. Something
+            #      like IoU between new and old (invalid) polygon would be
+            #      better, but can probably only be computed for pairs of valid
+            #      polygons. Maybe something based on pointwise distances,
+            #      where the points are sampled on the edges (not edge vertices
+            #      themselves). Maybe something based on drawing the perimeter
+            #      on images or based on distance maps.
+            point_kept_idx_to_pos = {point_idx: i for i, point_idx in enumerate(points_kept)}
+
+            # generate all possible combinations from <points_kept> and <points_left>
+            combos = np.transpose([np.tile(np.int32(points_left), len(np.int32(points_kept))),
+                                   np.repeat(np.int32(points_kept), len(np.int32(points_left)))])
+            combos = np.concatenate(
+                (combos, np.zeros((combos.shape[0], 3), dtype=np.int32)),
+                axis=1)
+
+            # copy columns 0, 1 into 2, 3 so that 2 is always the lower value
+            mask = combos[:, 0] < combos[:, 1]
+            combos[:, 2:4] = combos[:, 0:2]
+            combos[mask, 2] = combos[mask, 1]
+            combos[mask, 3] = combos[mask, 0]
+
+            # distance (in indices) between each pair of <point_kept> and <point_left>
+            combos[:, 4] = np.minimum(
+                combos[:, 3] - combos[:, 2],
+                len(points) - combos[:, 3] + combos[:, 2]
+            )
+
+            # limit candidates
+            max_dist = self.fit_max_dist_other_iters
+            if iteration > 0:
+                max_dist = self.fit_max_dist_first_iter
+            candidate_rows = combos[combos[:, 4] <= max_dist]
+            if self.fit_n_candidates_before_sort_max is not None \
+                    and len(candidate_rows) > self.fit_n_candidates_before_sort_max:
+                random_state.shuffle(candidate_rows)
+                candidate_rows = candidate_rows[0:self.fit_n_candidates_before_sort_max]
+
+            for row in candidate_rows:
+                point_left_idx = row[0]
+                point_kept_idx = row[1]
+                in_points_kept_pos = point_kept_idx_to_pos[point_kept_idx]
+                segment_start_idx = point_kept_idx
+                segment_end_idx = points_kept[(in_points_kept_pos+1) % len(points_kept)]
+                segment_start = points[segment_start_idx]
+                segment_end = points[segment_end_idx]
+                if iteration == 0:
+                    dist_eucl = 0
+                else:
+                    dist_eucl = _compute_distance_point_to_line(
+                        points[point_left_idx], segment_start, segment_end)
+                candidates.append((point_left_idx, point_kept_idx, row[4], dist_eucl))
+
+            # Sort computed distances first by minimal vertex-distance (see
+            # above, metric 1) (ASC), then by euclidean distance
+            # (metric 2) (ASC).
+            candidate_ids = np.arange(len(candidates))
+            candidate_ids = sorted(candidate_ids, key=lambda idx: (candidates[idx][2], candidates[idx][3]))
+            if self.fit_n_changes_max is not None:
+                candidate_ids = candidate_ids[:self.fit_n_changes_max]
+
+            # Iterate over point-segment pairs in sorted order. For each such
+            # candidate: Add the point to the already collected points,
+            # create a polygon from that and check if the polygon is valid.
+            # If it is, add the point to the output list and recalculate
+            # distance metrics. If it isn't valid, proceed with the next
+            # candidate until no more candidates are left.
+            #
+            # small change: this now no longer breaks upon the first found point
+            # that leads to a valid polygon, but checks all candidates instead
+            is_valid = False
+            done = set()
+            for candidate_idx in candidate_ids:
+                point_left_idx = candidates[candidate_idx][0]
+                point_kept_idx = candidates[candidate_idx][1]
+                if (point_left_idx, point_kept_idx) not in done:
+                    in_points_kept_idx = [i for i, point_idx in enumerate(points_kept) if point_idx == point_kept_idx][0]
+                    points_kept_hypothesis = points_kept[:]
+                    points_kept_hypothesis.insert(in_points_kept_idx+1, point_left_idx)
+                    poly_hypothesis = Polygon([points[idx] for idx in points_kept_hypothesis])
+                    if poly_hypothesis.is_valid:
+                        is_valid = True
+                        points_kept = points_kept_hypothesis
+                        points_left = [point_idx for point_idx in points_left if point_idx != point_left_idx]
+                        n_changes += 1
+                        if n_changes >= self.fit_n_changes_max:
+                            return points_kept
+                    done.add((point_left_idx, point_kept_idx))
+                    done.add((point_kept_idx, point_left_idx))
+
+            # none of the left points could be used to create a valid polygon?
+            # (this automatically covers the case of no points being left)
+            if not is_valid and iteration > 0:
+                converged = True
+
+            iteration += 1
+            if self.fit_n_iters_max is not None and iteration > self.fit_n_iters_max:
+                break
+
+        return points_kept
 
 
 class MultiPolygon(object):